==> samples/01-fake-news-classification-isot-dataset-cleaning.ipynb <==
import pandas as pd

from google.colab import drive
drive.mount('/content/gdrive')

IMPORT THE DATASETS

path_fake_news = '/content/gdrive/MyDrive/ai-playground/ra-text-classification-spectrogram/datasets/Fake.csv'
path_true_news = '/content/gdrive/MyDrive/ai-playground/ra-text-classification-spectrogram/datasets/True.csv'

df_fake_news = pd.read_csv(path_fake_news)
df_true_news = pd.read_csv(path_true_news)

df_fake_news.head()

df_true_news.head()

CATEGORIZE AND MERGE THE DATASETS

df_fake_news['category'] = 0
df_true_news['category'] = 1

df = pd.concat([df_true_news, df_fake_news])

MERGE ALL THE RELEVANT TEXT </br>
DELETE IRRELEVANT COLUMNS

df['text'] = df['text'] + ' ' + df['title']

del df['title']
del df['subject']
del df['date']

PRE-PROCESS THE TEXT </br>
<p> Remove URLs </p>

import re

def remove_urls(text):
  return re.sub(r'https\S+', '', text)

df['text'] = df['text'].apply(remove_urls)

<p> Strip HTML </p>

from bs4 import BeautifulSoup

def strip_html(text):
  soup = BeautifulSoup(text, 'html.parser')
  return soup.get_text()

df['text'] = df['text'].apply(strip_html)

<p> Remove square brackets </p>

def remove_between_square_brackets(text):
    return re.sub('\[[^]]*\]', '', text)

df['text'] = df['text'].apply(remove_between_square_brackets)

<p> Remove stopwords and punctuations </p>

import nltk
nltk.download('stopwords')

from nltk.corpus import stopwords

import string
from string import punctuation

stop = set(stopwords.words('english'))
punctuation = list(string.punctuation)
stop.update(punctuation)

def remove_stopwords(text):
  final_text = []
  for i in text.split():
      if i.strip().lower() not in stop:
          final_text.append(i.strip())
  return " ".join(final_text)

df['text'] = df['text'].apply(remove_stopwords)

df.to_csv('isot-dataset-cleaned.csv')

!cp isot-dataset-cleaned.csv "/content/gdrive/MyDrive/ai-playground/ra-text-classification-spectrogram/datasets"
