==> samples/02-isot-dataset-4000-words.ipynb <==
from google.colab import drive
drive.mount('/content/gdrive')

path_dataset = '/content/gdrive/MyDrive/ai-playground/ra-text-classification-spectrogram/datasets/isot-dataset-cleaned.csv'

import pandas as pd

df = pd.read_csv(path_dataset)

mask = (df['text'].str.len() >= 4000)
df4000 = df.loc[mask]

df4000.shape

len(df4000.loc[df['category'] == 1])

len(df4000.loc[df['category'] == 0])

df4000.to_csv('isot-dataset-cleaned-4000.csv')
!cp isot-dataset-cleaned-4000.csv "/content/gdrive/MyDrive/ai-playground/ra-text-classification-spectrogram/datasets"
