==> samples/03-isot-dataset-4000-true-BERT-embeddings.ipynb <==
from google.colab import drive
drive.mount('/content/gdrive')

path_dataset = '/content/gdrive/MyDrive/ai-playground/ra-text-classification-spectrogram/datasets/isot-dataset-cleaned-4000.csv'

import pandas as pd

df = pd.read_csv(path_dataset)

df_true_news = df.loc[df['category'] == 1]
df_true_news.shape[0]

!pip install simpletransformers

df_true_news_list = df_true_news['text'].tolist()
df_true_news_list[:5]
